Test fixes and refactoring in endurance tests
