+++ src/X509.EnduranceTest.Shared/TestMain.cs
@@ -14,13 +14,10 @@ namespace X509.EnduranceTest.Shared
 {
     public static class TestMain
     {
-        //...............................................................................
         #region AppSettings
-        //...............................................................................
-
-        static string   X509Thumbprint      => AppSetting("X509.Thumbprint");
-        static int      SampleDataSizeKB    => Convert.ToInt32(AppSetting("SampleDataSizeKB"));
-        static int      LoopCount           => Convert.ToInt32(AppSetting("LoopCount"));
+        
+        static int SampleDataSizeKB => Convert.ToInt32(AppSetting("SampleDataSizeKB"));
+        static int LoopCount => Convert.ToInt32(AppSetting("LoopCount"));
 
         static string AppSetting(string name) => ConfigurationManager.AppSettings[name] ?? throw new Exception($"AppSetting 'name' not defined.");
 
@@ -34,16 +31,13 @@ namespace X509.EnduranceTest.Shared
             }
             catch (Exception err)
             {
-
                 var topError = err;
-
                 while (null != err)
                 {
                     Console.WriteLine($"[{err.GetType().FullName}]");
                     Console.WriteLine(err.Message);
                     err = err.InnerException;
                 }
-
                 Console.WriteLine(topError.StackTrace);
             }
         }
@@ -79,7 +73,7 @@ namespace X509.EnduranceTest.Shared
                         break;
 
                     case "P":
-                        PrintCSP(cert);
+                        ConsoleWriter.PrintCSP(cert);
                         break;
 
                     case "Q":
@@ -100,8 +94,8 @@ namespace X509.EnduranceTest.Shared
             Console.WriteLine($"Generated {sampleData.Length / 1024} KB random binary data.");
 
             // Encrypt/Decrypt ONCE...
-            var encryptedBytes = EncryptBytes(sampleData, cert);
-            var decryptedBytes = DecryptBytes(encryptedBytes, cert);
+            var encryptedBytes = EncryptionDecryptionUtils.EncryptBytesUsingExtensionMethod(cert, sampleData);
+            var decryptedBytes = EncryptionDecryptionUtils.DecryptBytesUsingExtensionMethod(cert, encryptedBytes);
             Console.WriteLine($"SampleData: {sampleData.Length:#,0} bytes");
             Console.WriteLine($"Encrypted:  {encryptedBytes.Length:#,0} bytes");
             Console.WriteLine($"Decrypted:  {decryptedBytes.Length:#,0} bytes");
@@ -111,72 +105,6 @@ namespace X509.EnduranceTest.Shared
             if (!good) throw new Exception("Decrypted result doesn't match original data.");
         }
 
-        static void PrintCSP(X509Certificate2 cert)
-        {
-            Console.WriteLine($"Cert: {cert.Subject} / {cert.Thumbprint}");
-            Console.WriteLine();
-
-            try
-            {
-                Console.WriteLine("cert.PublicKey.Key");
-                var alg = cert.PublicKey.Key;
-                Console.WriteLine(alg.GetType().FullName);
-                Console.WriteLine();
-            }
-            catch (Exception err)
-            {
-                PrintErrorSummary(err);
-            }
-
-            try
-            {
-                // Fails in .NET Framework if -KeySpec Signature not specified.
-                // Works in .NET Core
-                Console.WriteLine("cert.PrivateKey");
-                var alg = cert.PrivateKey;
-                Console.WriteLine(alg.GetType().FullName);
-                Console.WriteLine();
-            }
-            catch (Exception err)
-            {
-                PrintErrorSummary(err);
-            }
-
-            try
-            {
-                Console.WriteLine("cert.GetRSAPublicKey()");
-                var alg = cert.GetRSAPublicKey();
-                Console.WriteLine(alg.GetType().FullName);
-                Console.WriteLine();
-            }
-            catch (Exception err)
-            {
-                PrintErrorSummary(err);
-            }
-
-            try
-            {
-                Console.WriteLine("cert.GetRSAPrivateKey()");
-                var alg = cert.GetRSAPrivateKey();
-                Console.WriteLine(alg.GetType().FullName);
-                Console.WriteLine();
-            }
-            catch (Exception err)
-            {
-                PrintErrorSummary(err);
-            }
-
-            void PrintErrorSummary(Exception ex)
-            {
-                Console.WriteLine("ERROR:");
-                while(null != ex)
-                {
-                    Console.WriteLine($"[{ex.GetType().FullName}]");
-                    Console.WriteLine(ex.Message);
-                    ex = ex.InnerException;
-                }
-            }
-        }
 
         static void BeginEncryptionLoop(X509Certificate2 cert, int maxIterations)
         {
@@ -195,8 +123,8 @@ namespace X509.EnduranceTest.Shared
             Console.WriteLine($"MaxIterations: {maxIterations:#,0}");
             Console.WriteLine($"Generated {sampleData.Length / 1024} KB random binary data.");
 
-            var encryptedBytes = EncryptBytes(sampleData, cert);
-            var decryptedBytes = DecryptBytes(encryptedBytes, cert);
+            var encryptedBytes = EncryptionDecryptionUtils.EncryptBytesUsingExtensionMethod(cert, sampleData);
+            var decryptedBytes = EncryptionDecryptionUtils.DecryptBytesUsingExtensionMethod(cert, encryptedBytes);
 
             var counter = 0;
             var elapsed = Stopwatch.StartNew();
@@ -207,8 +135,8 @@ namespace X509.EnduranceTest.Shared
 
             while (counter++ <= maxIterations)
             {
-                if (encrypt) EncryptBytes(decryptedBytes, cert);
-                if (decrypt) DecryptBytes(encryptedBytes, cert);
+                if (encrypt) EncryptionDecryptionUtils.EncryptBytesUsingExtensionMethod(cert, decryptedBytes);
+                if (decrypt) EncryptionDecryptionUtils.DecryptBytesUsingExtensionMethod(cert, encryptedBytes);
 
                 if (nextStatusUpdate < DateTime.Now)
                 {
@@ -217,34 +145,11 @@ namespace X509.EnduranceTest.Shared
                     nextStatusUpdate = DateTime.Now.Add(statusUpdateInterval);
                 }
             }
-
             rate = counter / elapsed.Elapsed.TotalSeconds;
             Console.WriteLine("Finished.");
             Console.WriteLine($"{elapsed.Elapsed:hh\\:mm\\:ss} @ {rate:#,0} per-sec. Iterations: {counter:#,0} (Use Ctrl-C to quit...)");
         }
-
-        static byte[] EncryptBytes(byte[] inputData, X509Certificate2 cert)
-        {
-            using (var input = new MemoryStream(inputData))
-            using (var output = new MemoryStream())
-            {
-                cert.EncryptStream(input, output);
-                output.Flush();
-                return output.ToArray();
-            }
-        }
-
-        static byte[] DecryptBytes(byte[] inputData, X509Certificate2 cert)
-        {
-            using (var input = new MemoryStream(inputData))
-            using (var output = new MemoryStream(inputData.Length))
-            {
-                cert.DecryptStream(input, output);
-                output.Flush();
-                return output.ToArray();
-            }
-        }
-
-       
+    }
+    public class EnduranceTestResult {
     }
 }
